# internal/interface/http/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: Stock Service API
  version: "1.0.0"
  description: API for managing inventory, stores, promotions, warehouses, and related resources.
paths:
...
  /api/v1/products:
    post:
      summary: Create a new product
      tags: [Product]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title_fa, brand_id, category_id]
              properties:
                title_fa:          { type: string }
                title_en:          { type: string }
                description:       { type: string }
                brand_id:          { type: integer, format: int64 }
                category_id:       { type: integer, format: int64 }
                owner_type:        { type: string, enum: [system, user] }
                owner_id:          { type: integer, format: int64, nullable: true }
                is_original:       { type: boolean }
                index_image_file_id: { type: integer, format: int64, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/products/{id}:
    get:
      summary: Get product by ID
      tags: [Product]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int32 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a product
      tags: [Product]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int32 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title_fa:          { type: string }
                title_en:          { type: string }
                description:       { type: string }
                brand_id:          { type: integer, format: int64 }
                category_id:       { type: integer, format: int64 }
                index_image_file_id: { type: integer, format: int64, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Soft delete a product
      tags: [Product]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int32 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/products/{id}/activate:
    post:
      summary: Activate a product
      tags: [Product]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int32 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/products/{id}/reject:
    post:
      summary: Reject a product
      tags: [Product]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int32 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/sales-commissions:
    post:
      summary: Create a sales commission
      tags: [Sales Commission]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inventory_id, category_commission_rule_id, sale_model, rate_percent, min_price]
              properties:
                inventory_id:                { type: integer, format: int64 }
                category_commission_rule_id: { type: integer, format: int64 }
                sale_model:                  { type: string }
                rate_percent:                { type: number, format: double }
                min_price:                   { type: number, format: double }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesCommissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sales-commissions/{id}/max-price:
    put:
      summary: Update max price for a sales commission
      tags: [Sales Commission]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [max_price]
              properties:
                max_price: { type: number, format: double }
      responses:
        "200":
          description: OK
  /api/v1/sales-commissions/{id}/min-qty:
    put:
      summary: Update min quantity for a sales commission
      tags: [Sales Commission]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [min_qty]
              properties:
                min_qty: { type: integer }
      responses:
        "200":
          description: OK
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title_fa:
          type: string
        title_en:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        brand_id:
          type: integer
          format: int64
        category_id:
          type: integer
          format: int64
        owner_type:
          type: string
        owner_id:
          type: integer
          format: int64
          nullable: true
        is_original:
          type: boolean
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        index_image_file_id:
          type: integer
          format: int64
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    SalesCommissionResponse:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        InventoryID:
          type: integer
          format: int64
        CategoryCommissionRuleID:
          type: integer
          format: int64
        SaleModel:
          type: string
        RatePercent:
          type: number
          format: double
        MinQty:
          type: integer
          nullable: true
        MinPrice:
          type: number
          format: double
        MaxPrice:
          type: number
          format: double
          nullable: true
        CreatedAt:
          type: string
          format: date-time
...